docs(examples): replace node-http with live-mount FUSE demo
Drop the older node-http example in favor of live-mount, which shows
off the bidirectional FUSE-backed liveMounts API. The host's ./shared/
appears inside the guest at /mnt/shared; edits propagate both ways
immediately, with no rebake or reboot.

run.ts boots with liveMounts + portForward, the guest serves
shared/server.mjs over HTTP, and the README walks through editing
message.txt on the host and tailing access.log written from the guest.

The node-http example covered the same boot+serve flow without the
mount-server angle; live-mount supersedes it.
