+++ examples/live-mount/.gitignore
@@ -1,0 +1,2 @@
+.cache/
+shared/access.log
