+++ examples/live-mount/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@machinen/example-node-http",
+  "name": "@machinen/example-live-mount",
   "version": "0.0.0",
   "private": true,
   "type": "module",
