+++ examples/live-mount/shared/message.txt
@@ -1,0 +1,1 @@
+hello from the host
